Fix issue with creative inventory tabs
Due to new features located in the creative inventory, the older layout is deprecated.
The texture file does not align with the new layout, causing weird visuel effects.

By reducing the width of each tab by 4px, we can fix the issue

Commit also changes the bottom of the tab